+++ CHANGELOG.rst
@@ -29,7 +29,7 @@ Changed
 - Restructured the way sources, documentation etc. are stored.
 - Added a rules for contributing to the project (see ``CONTRIBUTING.md``).
 - include and man files are installed
-- Added a script to make testing simpler (``scripts/run_test``)
+- Added a script to make testing simpler (``scripts/run_test.py``)
 
 Deprecated
 ==========